refactor: pull CardButton into Cards component folder
- also pulling in some of the styles from the Settings page
