+++ packages/widget/src/components/Card/index.ts
@@ -3,3 +3,5 @@ export * from './CardHeader';
 export * from './CardIconButton';
 export * from './CardLabel';
 export * from './CardTitle';
+export * from './CardButton';
+export * from './Card.style';
