+++ packages/widget/src/pages/SettingsPage/LanguageSetting.tsx
@@ -5,8 +5,7 @@ import { useLanguages } from '../../hooks';
 import { useWidgetConfig } from '../../providers';
 import { HiddenUI } from '../../types';
 import { navigationRoutes } from '../../utils';
-import { CardButton } from '../../components/CardButton';
-import { SummaryValue } from './SettingsCard';
+import { CardButton, CardValue } from '../../components/Card';
 
 export const LanguageSetting: React.FC = () => {
   const { t } = useTranslation();
@@ -28,7 +27,7 @@ export const LanguageSetting: React.FC = () => {
       icon={<LanguageIcon />}
       title={t(`language.title`)}
     >
-      <SummaryValue>{selectedLanguageDisplayName}</SummaryValue>
+      <CardValue>{selectedLanguageDisplayName}</CardValue>
     </CardButton>
   );
 };
