--- packages/widget/src/pages/SettingsPage/SettingsCard/SettingCard.tsx
@@ -1,16 +1,0 @@
-import type { PropsWithChildren } from 'react';
-import { Card } from '../../../components/Card';
-
-export const SettingCard: React.FC<PropsWithChildren> = ({ children }) => {
-  return (
-    <Card
-      variant="default"
-      selectionColor="primary"
-      sx={{
-        p: 1,
-      }}
-    >
-      {children}
-    </Card>
-  );
-};
