+++ packages/widget/src/pages/SettingsPage/SettingsCard/index.ts
@@ -1,4 +1,3 @@
-export * from './SettingCard';
 export * from './SettingCard.style';
 export * from './SettingCardExpandable';
 export * from './BadgedValue';
